Reworked repository base.
